+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Mouse wheel support for sliders.
-- Control-click to set sliders to default value.
+- Control-click or right-click to set a slider to its default value.
 - Slider fine control by holding shift when dragging.
 - Strip symbols from release builds for smaller binaries.
 - Disable buttons when device is not connected.